+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/model/widgets/ExplorerNode.java
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,11 +22,7 @@
 package com.intellij.idea.plugin.hybris.system.cockpitng.model.widgets;
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants;
-import com.intellij.util.xml.DomElement;
-import com.intellij.util.xml.GenericAttributeValue;
-import com.intellij.util.xml.Namespace;
-import com.intellij.util.xml.Required;
-import com.intellij.util.xml.SubTagList;
+import com.intellij.util.xml.*;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -36,6 +32,7 @@ import org.jetbrains.annotations.NotNull;
 public interface ExplorerNode extends DomElement, Positioned {
 
     String ID = "id";
+    String MERGE_MODE = "merge-mode";
 
     /**
      * Returns the value of the id child.
@@ -64,7 +61,7 @@ public interface ExplorerNode extends DomElement, Positioned {
      * @return the value of the merge-mode child.
      */
     @NotNull
-    @com.intellij.util.xml.Attribute("merge-mode")
+    @com.intellij.util.xml.Attribute(MERGE_MODE)
     GenericAttributeValue<String> getMergeMode();